backend/requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.26.4
mysql-connector-python==8.2.0
scikit-learn>=1.4.0